requirements.txt
dateparser==0.7.6
jinja2==2.11.3
oauthlib==3.1.0
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
markupsafe==2.0.1
